Container that can run integration tests in Montagu deploy environment
