+++ app/.dockerignore
@@ -1,3 +1,4 @@
+.tmp
 node_modules
 npm-debug.log
 src/main/contrib/models/Generated.ts
