+++ app/docker/integration-tests.dockerfile
@@ -1,5 +1,7 @@
 FROM montagu-portal-build-env
 
-RUN npm install -g mocha-webpack
-RUN webpack
-CMD ./scripts/run-integration-tests.sh
+ENV PGHOST db
+ENV MONTAGU_PORTAL_PROFILE packaged_integration_tests
+
+CMD ./scripts/add-test-accounts-for-integration-tests.sh && \
+    ./scripts/run-integration-tests.sh
