+++ app/scripts/run-integration-tests-with-apis.sh
@@ -3,6 +3,10 @@ set -ex
 
 export here=$(dirname $0)
 source $here/run-apis.sh
+
+# Tell the tests where to find the database
+export PGHOST=localhost
+
 $here/run-integration-tests.sh "$@"
 result=$?
