+++ app/scripts/run-integration-tests.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 set -ex
+# Note: PGHOST environment variable must be set; see below
 
 # Either specify a portal with "Admin" or "Contrib", or fall back to *, which 
 # will run all tests
@@ -8,11 +9,13 @@ export portal=${1:-*}
 # Set database variables (the names here are used by the Node pg library,
 # and shouldn't be changed)
 export PGUSER=vimc
-export PGHOST=localhost
 export PGPASSWORD=changeme
 export PGDATABASE=montagu
 export PGTEMPLATE=montagu_template
 export PGPORT=5432
+# We also need PGHOST to be set, but this varies based on whether the tests are
+# being run in a container or not, so we require that to be set by the calling
+# script
 
 # Setup template database
 docker exec montagu_db_1 psql -U vimc -d postgres -c \
