+++ app/src/main/shared/Settings.ts
@@ -1,7 +1,8 @@
+import { settings as defaultSettings } from "./settings/default";
 const environment = require('environmentSettings');
 const appName = require('appName');
 
-export const settings: Settings = environment.settings;
+export const settings: Settings = Object.assign({}, defaultSettings, environment.settings);
 export const appSettings: AppSpecificSettings = (settings as any)[appName.appName];
 
 console.log("Current settings: " + JSON.stringify(settings));
