+++ app/src/main/shared/settings/teamcity.ts
@@ -1,19 +1,3 @@
-export const settings: Settings = {
-    reportingApiUrl: () => "http://localhost:8081/v1",
-    apiUrl: () => "http://localhost:8080/v1",
-    montaguUrl: () => "http://localhost:5000/",
-    supportContact: "montagu-help@imperial.ac.uk",
+export const settings: Partial<Settings> = {
     teamcityServiceMessages: true,
-    admin: {
-        publicPath: "",
-        requiresModellingGroupMembership: false
-    },
-    contrib: {
-        publicPath: "",
-        requiresModellingGroupMembership: true
-    },
-    report: {
-        publicPath: "",
-        requiresModellingGroupMembership: false
-    }
 };
